DICOM file restructure
