+++ src/Model/DICOM/DICOMWidgetItem.py
