+++ test/test_model_batch_processes.py
@@ -7,9 +7,7 @@ from pydicom import dcmread
 from pydicom.errors import InvalidDicomError
 from PySide6.QtWidgets import QApplication
 from src.Controller.BatchProcessingController import BatchProcessingController
-from src.Model import DICOMDirectorySearch
-from src.Model import DICOMStructuredReport
-from src.Model import ROI
+from src.Model.DICOM import DICOMDirectorySearch, DICOMStructuredReport
 from src.Model.batchprocessing.BatchProcessClinicalDataSR2CSV import \
     BatchProcessClinicalDataSR2CSV
 from src.Model.batchprocessing.BatchProcessCSV2ClinicalDataSR import \
